requirements.txt
langdetect==1.0.7
Pillow==10.3.0
PyICU==2.13.0
requests==2.22.0
